Improved code structure and some fixes
