--- core/app/[locale]/(default)/(faceted)/_components/attribute-labels.json
@@ -1,15 +1,0 @@
-{
-  "categories.lvl0": "Category",
-  "categories.lvl1": "Category",
-  "categories.lvl2": "Category",
-  "brand_name": "Brand",
-  "metafields.Akeneo.collection": "Collection",
-  "metafields.Akeneo.number_of_bulbs": "Number of Lights",
-  "variants.options.Finish Color": "Finish Color",
-  "variants.options.Shade Color": "Shade Color",
-  "metafields.Akeneo.product_style": "Style",
-  "metafields.Akeneo.product_shape": "Shape",
-  "prices.USD": "Price",
-  "is_new": "Is New",
-  "on_sale": "On Sale"
-}
