+++ core/app/[locale]/(default)/(faceted)/brand/[slug]/page.tsx
@@ -2,40 +2,15 @@ import type { Metadata } from 'next';
 import { notFound } from 'next/navigation';
 import { getTranslations, setRequestLocale } from 'next-intl/server';
 
-import { ProductCard } from '~/components/product-card';
-import { Pagination } from '~/components/ui/pagination';
-
-import { FacetedSearch } from '../../_components/faceted-search';
-import { MobileSideNav } from '../../_components/mobile-side-nav';
-import { SortBy } from '../../_components/sort-by';
-import { fetchFacetedSearch } from '../../fetch-faceted-search';
-
 import { getSessionCustomerAccessToken } from '~/auth';
 
 import { Breadcrumbs } from '~/components/breadcrumbs';
 
 import Link from 'next/link';
-import Image from 'next/image';
-import logo from '~/public/brand/logo.svg';
-
-import image1_1 from '~/public/brand/image1-1.svg';
-import image1_2 from '~/public/brand/image1-2.svg';
-import image1_3 from '~/public/brand/image1-3.svg';
-import image1_4 from '~/public/brand/image1-4.svg';
-import image1_5 from '~/public/brand/image1-5.svg';
-
-import image2_1 from '~/public/brand/image2-1.svg';
-import image2_2 from '~/public/brand/image2-2.svg';
-import image2_3 from '~/public/brand/image2-3.svg';
-import image2_4 from '~/public/brand/image2-4.svg';
-import image2_5 from '~/public/brand/image2-5.svg';
-
-import image1 from '~/public/brand/image1.svg';
-import image2 from '~/public/brand/image2.svg';
 
 import { getBrand } from './page-data';
 
-import { getPromotions } from '../../fetch-promotions';
+import { getPromotions } from '~/belami/lib/fetch-promotions';
 
 import { Brand } from './brand';
 
@@ -98,7 +73,6 @@ export default async function BrandPage(props: Props) {
   });
 
   const promotions = await getPromotions();
-  const hasExtendedContent = false;
 
   return (
     <div className="group py-4 px-4 xl:px-12">
@@ -158,367 +132,6 @@ export default async function BrandPage(props: Props) {
       {!!snapshot &&
         <MakeswiftPage snapshot={snapshot} />
       }
-
-      {hasExtendedContent && (
-        <>
-          <section className="mt-16">
-            <h3 className="text-lg font-medium">Section Title</h3>
-
-            <div className="relative mt-4 w-full overflow-x-hidden">
-              <ul className="flex space-x-4">
-                <li key="1" className="w-60">
-                  <Image src={image1_1} alt="Image 1" className="w-full" />
-                  <div className="mt-2 flex items-center justify-center space-x-2">
-                    <span className="uppercase">Category Name</span>
-                    <svg
-                      width="8"
-                      height="13"
-                      viewBox="0 0 8 13"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                        fill="#1C1B1F"
-                      />
-                    </svg>
-                  </div>
-                </li>
-                <li key="2" className="w-60">
-                  <Image src={image1_2} alt="Image 2" className="w-full" />
-                  <div className="mt-2 flex items-center justify-center space-x-2">
-                    <span className="uppercase">Category Name</span>
-                    <svg
-                      width="8"
-                      height="13"
-                      viewBox="0 0 8 13"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                        fill="#1C1B1F"
-                      />
-                    </svg>
-                  </div>
-                </li>
-                <li key="3" className="w-60">
-                  <Image src={image1_3} alt="Image 3" className="w-full" />
-                  <div className="mt-2 flex items-center justify-center space-x-2">
-                    <span className="uppercase">Category Name</span>
-                    <svg
-                      width="8"
-                      height="13"
-                      viewBox="0 0 8 13"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                        fill="#1C1B1F"
-                      />
-                    </svg>
-                  </div>
-                </li>
-                <li key="4" className="w-60">
-                  <Image src={image1_4} alt="Image 4" className="w-full" />
-                  <div className="mt-2 flex items-center justify-center space-x-2">
-                    <span className="uppercase">Category Name</span>
-                    <svg
-                      width="8"
-                      height="13"
-                      viewBox="0 0 8 13"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                        fill="#1C1B1F"
-                      />
-                    </svg>
-                  </div>
-                </li>
-                <li key="5" className="w-60">
-                  <Image src={image1_5} alt="Image 5" className="w-full" />
-                  <div className="mt-2 flex items-center justify-center space-x-2">
-                    <span className="uppercase">Category Name</span>
-                    <svg
-                      width="8"
-                      height="13"
-                      viewBox="0 0 8 13"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                        fill="#1C1B1F"
-                      />
-                    </svg>
-                  </div>
-                </li>
-                <li key="6" className="w-60">
-                  <Image src={image1_1} alt="Image 1" className="w-full" />
-                  <div className="mt-2 flex items-center justify-center space-x-2">
-                    <span className="uppercase">Category Name</span>
-                    <svg
-                      width="8"
-                      height="13"
-                      viewBox="0 0 8 13"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                        fill="#1C1B1F"
-                      />
-                    </svg>
-                  </div>
-                </li>
-                <li key="7" className="w-60">
-                  <Image src={image1_2} alt="Image 2" className="w-full" />
-                  <div className="mt-2 flex items-center justify-center space-x-2">
-                    <span className="uppercase">Category Name</span>
-                    <svg
-                      width="8"
-                      height="13"
-                      viewBox="0 0 8 13"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                        fill="#1C1B1F"
-                      />
-                    </svg>
-                  </div>
-                </li>
-              </ul>
-              <button
-                type="button"
-                title="Next"
-                className="absolute right-4 top-1/2 -mt-6 flex h-12 w-12 items-center justify-center rounded-full bg-brand-600 text-white"
-              >
-                <svg
-                  width="8"
-                  height="13"
-                  viewBox="0 0 8 13"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                    fill="currentColor"
-                  />
-                </svg>
-              </button>
-            </div>
-          </section>
-
-          <section className="mt-16">
-            <h3 className="text-lg font-medium">Section Title</h3>
-
-            <div className="relative mt-4 w-full overflow-x-hidden">
-              <ul className="flex space-x-4">
-                <li key="1" className="w-60">
-                  <Image src={image2_1} alt="Image 1" className="w-full" />
-                  <div className="mt-2 flex items-center justify-center space-x-2">
-                    <span className="uppercase">Category Name</span>
-                    <svg
-                      width="8"
-                      height="13"
-                      viewBox="0 0 8 13"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                        fill="#1C1B1F"
-                      />
-                    </svg>
-                  </div>
-                </li>
-                <li key="2" className="w-60">
-                  <Image src={image2_2} alt="Image 2" className="w-full" />
-                  <div className="mt-2 flex items-center justify-center space-x-2">
-                    <span className="uppercase">Category Name</span>
-                    <svg
-                      width="8"
-                      height="13"
-                      viewBox="0 0 8 13"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                        fill="#1C1B1F"
-                      />
-                    </svg>
-                  </div>
-                </li>
-                <li key="3" className="w-60">
-                  <Image src={image2_3} alt="Image 3" className="w-full" />
-                  <div className="mt-2 flex items-center justify-center space-x-2">
-                    <span className="uppercase">Category Name</span>
-                    <svg
-                      width="8"
-                      height="13"
-                      viewBox="0 0 8 13"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                        fill="#1C1B1F"
-                      />
-                    </svg>
-                  </div>
-                </li>
-                <li key="4" className="w-60">
-                  <Image src={image2_4} alt="Image 4" className="w-full" />
-                  <div className="mt-2 flex items-center justify-center space-x-2">
-                    <span className="uppercase">Category Name</span>
-                    <svg
-                      width="8"
-                      height="13"
-                      viewBox="0 0 8 13"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                        fill="#1C1B1F"
-                      />
-                    </svg>
-                  </div>
-                </li>
-                <li key="5" className="w-60">
-                  <Image src={image2_5} alt="Image 5" className="w-full" />
-                  <div className="mt-2 flex items-center justify-center space-x-2">
-                    <span className="uppercase">Category Name</span>
-                    <svg
-                      width="8"
-                      height="13"
-                      viewBox="0 0 8 13"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                        fill="#1C1B1F"
-                      />
-                    </svg>
-                  </div>
-                </li>
-                <li key="6" className="w-60">
-                  <Image src={image2_1} alt="Image 1" className="w-full" />
-                  <div className="mt-2 flex items-center justify-center space-x-2">
-                    <span className="uppercase">Category Name</span>
-                    <svg
-                      width="8"
-                      height="13"
-                      viewBox="0 0 8 13"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                        fill="#1C1B1F"
-                      />
-                    </svg>
-                  </div>
-                </li>
-                <li key="7" className="w-60">
-                  <Image src={image2_2} alt="Image 2" className="w-full" />
-                  <div className="mt-2 flex items-center justify-center space-x-2">
-                    <span className="uppercase">Category Name</span>
-                    <svg
-                      width="8"
-                      height="13"
-                      viewBox="0 0 8 13"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                        fill="#1C1B1F"
-                      />
-                    </svg>
-                  </div>
-                </li>
-              </ul>
-              <button
-                type="button"
-                title="Next"
-                className="absolute right-4 top-1/2 -mt-6 flex h-12 w-12 items-center justify-center rounded-full bg-brand-600 text-white"
-              >
-                <svg
-                  width="8"
-                  height="13"
-                  viewBox="0 0 8 13"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M4.90005 6.35089L0.300049 1.75089L1.70005 0.350891L7.70005 6.35089L1.70005 12.3509L0.300049 10.9509L4.90005 6.35089Z"
-                    fill="currentColor"
-                  />
-                </svg>
-              </button>
-            </div>
-          </section>
-
-          <section className="mt-16 flex h-48 items-center justify-center bg-gray-50 px-4 py-8 text-center">
-            <h3 className="text-lg font-medium">Promotional Content Placeholder</h3>
-          </section>
-
-          <section className="mt-16">
-            <h3 className="text-lg font-medium">More Ways to Shop</h3>
-
-            <div className="mt-4 flex w-full items-start space-x-4">
-              <div className="">
-                <Image src={image1} alt="Image 1" className="w-full" />
-                <h4 className="text-md mt-4 font-medium">A Title About This Item</h4>
-                <p className="mt-2">
-                  Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-                  incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
-                  exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute
-                  irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
-                  pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui
-                  officia deserunt mollit anim id est laborum.
-                </p>
-                <p className="mt-2">
-                  <a href="#" className="uppercase">
-                    Link Here
-                  </a>
-                </p>
-              </div>
-              <div className="">
-                <Image src={image2} alt="Image 2" className="w-full" />
-                <h4 className="text-md mt-4 font-medium">A Title About This Item</h4>
-                <p className="mt-2">
-                  Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-                  incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
-                  exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute
-                  irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
-                  pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui
-                  officia deserunt mollit anim id est laborum.
-                </p>
-                <p className="mt-2">
-                  <a href="#" className="uppercase">
-                    Link Here
-                  </a>
-                </p>
-              </div>
-            </div>
-          </section>
-
-          <section className="mt-16 flex h-48 items-center justify-center bg-gray-50 px-4 py-8 text-center">
-            <h3 className="text-lg font-medium">Promotional Content Placeholder</h3>
-          </section>
-
-          <h3 className="mt-16 text-lg font-medium">Products Catalog</h3>
-        </>
-      )}
 
       <Brand brand={brand} promotions={promotions} useDefaultPrices={useDefaultPrices} />
     </div>
