+++ core/app/[locale]/(default)/product/[slug]/collection-products.tsx
@@ -20,7 +20,7 @@ import { cn } from '~/lib/utils';
 
 import { ReviewSummary } from '~/belami/components/reviews';
 
-import searchColors from './search-colors.json';
+import searchColors from '~/belami/include/search-colors.json';
 
 type DynamicObject = {
   [key: string]: string;
