+++ core/app/[locale]/(default)/product/[slug]/page.tsx
@@ -26,7 +26,7 @@ import { RelatedProducts } from './related-products';
 import { CollectionProducts } from './collection-products';
 import { SitevibesReviews } from './sitevibes-reviews';
 
-import { getRelatedProducts, getCollectionProducts } from './fetch-algolia-products';
+import { getRelatedProducts, getCollectionProducts } from '../../../../../belami/lib/fetch-algolia-products';
 
 interface Props {
   params: Promise<{ slug: string; locale: string }>;
