--- core/app/[locale]/(default)/product/[slug]/search-colors.json
@@ -1,32 +1,0 @@
-{
-  "Off-White": "#efeae7",
-  "Bronze/Dark Brown": "#4d3832",
-  "Nickel/Silver": "#c9c9c9",
-  "Orange": "#d6913b",
-  "Brass": "#cab666",
-  "Copper": "#c07143",
-  "Pewter": "#7b7e83",
-  "Iron/Dark Gray": "#474849",
-  "Black": "#000000",
-  "Contrasting Finishes": "Contrasting Finishes.svg",
-  "Light Brown": "#b0875d",
-  "Blue": "#3a4ad6",
-  "Multi-Color": "Multi-Color.svg",
-  "Gold": "#ddb336",
-  "Chrome": "#a8a9ad",
-  "Clear": "#f4fafc",
-  "Wood": "#a1662f",
-  "Green": "#33d130",
-  "White": "#ffffff",
-  "Purple": "#964fd1",
-  "Red": "#cc0a30",
-  "Stone": "#b9b0a4",
-  "Galvanized": "#a8bdc6",
-  "Mixed Metals": "Mixed Metals.svg",
-  "Alabaster": "#edeade",
-  "Marble": "Marble.svg",
-  "Concrete": "#d2d1cd",
-  "Yellow": "#fffb72",
-  "Bisque": "#faf4ed",
-  "Pink": "#ffbce5"
-}
