+++ core/belami/include/attribute-labels.json
@@ -1,0 +1,66 @@
+{
+  "categories.lvl0": "Category",
+  "categories.lvl1": "Category",
+  "categories.lvl2": "Category",
+  "brand_name": "Brand",
+  "reviews_rating_sum": "Rating",
+  "metafields.Details.Collection": "Collection",
+  "metafields.Details.Number of Bulbs": "Number of Lights",
+  "metafields.Details.Product Style": "Style",
+  "metafields.Details.Fuel Source": "Fuel Source",
+  "metafields.Details.Voltage": "Voltage Rating",
+  "metafields.Details.Product Shape": "Shape",
+  "metafields.Details.Length": "Length",
+  "metafields.Details.Weight": "Weight",
+  "metafields.Details.Depth": "Depth",
+  "metafields.Details.Width": "Width",
+  "metafields.Details.Height": "Height",
+  "metafields.Details.Application": "Application",
+  "metafields.Details.Base Type": "Base Type",
+  "metafields.Details.Blade Span": "Blade Span",
+  "metafields.Details.Blade to Ceiling Height": "Blade to Ceiling Height",
+  "metafields.Details.Color Temperature": "Color Temperature",
+  "metafields.Details.Control Type": "Control Type",
+  "metafields.Details.Crystal Type": "Crystal Type",
+  "metafields.Details.Designer": "Designer",
+  "metafields.Details.Controls Included": "Controls Included",
+  "metafields.Details.Hanging Method": "Hanging Method",
+  "metafields.Details.Heat Rating": "Heat Rating",
+  "metafields.Details.Heating Area": "Heating Area",
+  "metafields.Details.IR Heat Technology": "IR Heat Technology",
+  "metafields.Details.Lift": "Lift",  
+  "metafields.Details.Lumen Output": "Lumen Output",
+  "metafields.Details.Material": "Material",  
+  "metafields.Details.Minimum Mounting Height": "Minimum Mounting Height",
+  "metafields.Details.Number of Blades": "Number of Blades",  
+  "metafields.Details.Number of Main Burners": "Number of Main Burners",
+  "metafields.Details.Photocell Included": "Photocell Included",
+  "metafields.Details.Pole Material": "Pole Material",
+  "metafields.Details.Pole Diameter": "Pole Diameter",
+  "metafields.Details.Power Source": "Power Source",
+  "metafields.Details.Product Design": "Product Design",
+  "metafields.Details.Shade Height": "Shade Height",
+  "metafields.Details.Shade Material": "Shade Material",
+  "metafields.Details.Shade Width Bottom": "Shade Width Bottom",
+  "metafields.Details.Shade Width Top": "Shade Width Top",
+  "metafields.Details.Smart Compatible": "Smart Compatible",
+  "metafields.Details.Smart Product": "Smart Product",
+  "metafields.Details.Switch Type": "Switch Type",
+  "metafields.Details.Tilt": "Tilt",
+  "metafields.Details.Total Grilling Area": "Total Grilling Area",
+  "metafields.Details.Voltage Options": "Voltage Options",
+  "metafields.Details.Wattage": "Wattage",
+  "metafields.Details.Can be Recessed": "Can be Recessed",
+  "variants.options.Finish Color": "Finish Color",
+  "variants.options.Glass Color": "Glass Color",
+  "variants.options.Shade Color": "Shade Color",
+  "variants.options.Blade Color": "Blade Color",
+  "metafields.Akeneo.product_style": "Style",
+  "metafields.Akeneo.product_shape": "Shape",
+  "prices.USD": "Price",
+  "is_new": "Is New",
+  "on_sale": "On Sale",
+  "on_clearance": "On Clearance",
+  "in_stock": "In Stock",
+  "free_shipping": "Free Shipping"
+}
