+++ core/belami/include/search-colors.json
