+++ core/belami/lib/fetch-algolia-products.ts
