+++ core/components/ui/header/header.tsx
@@ -62,7 +62,7 @@ const Header = ({
   homeLogoMobileFirst,
 }: Props) => {
   const [openMenuId, setOpenMenuId] = useState<string | null>(null);
-  const timeoutRef = useRef<NodeJS.Timeout>();
+  const timeoutRef = useRef<NodeJS.Timeout>(null);
 
   const handleMenuEnter = (linkHref: string) => {
     if (timeoutRef.current) {
@@ -111,7 +111,7 @@ const Header = ({
           </CustomLink>
 
           <CustomLink className="home-logo-two overflow-hidden text-ellipsis pt-3 md:pl-5" href="/">
-            {typeof logo === 'object' ? (
+            {typeof logo === 'object' && homeLogoMobileFirst ? (
               <div className="second-home-logo block lg:hidden">
                 <BcImage
                   alt="homeLogo"
@@ -119,7 +119,7 @@ const Header = ({
                   height={30}
                   priority
                   // src={imageIconList.homeLogo}
-                  src={homeLogoMobileFirst}
+                  src={homeLogoMobileFirst || ''}
                   width={30}
                   unoptimized={true}
                 />
@@ -180,6 +180,7 @@ const Header = ({
                         className={`header-sub-menu-${menuIndex} relative mx-auto grid max-w-[90em] gap-6 px-12 md:grid-cols-[repeat(2,auto)] lg:grid-cols-[repeat(2,auto)] xl:grid-cols-[repeat(6,auto)]`}
                       >
                         <button
+                          title="Close menu"
                           className="absolute -top-[1em] right-[1em] text-gray-600 hover:text-gray-800 focus:outline-none"
                           onClick={handleMenuClose}
                         >
